Extract shared routing infrastructure to src/routing/
Move graph, router, inserters, and common utilities from src/spaghetti/
into a new src/routing/ module shared by both layout engines:

- src/routing/graph.py — ProductionGraph, FlowEdge, MachineNode
- src/routing/router.py — A* pathfinding, route_connections
- src/routing/inserters.py — lane-aware inserter assignment
- src/routing/common.py — machine_size, machine_tiles, belt_entity_for_rate, DIR_MAP
- src/routing/orchestrate.py — shared build_layout (deduplicates _attempt_layout)

Spaghetti files become re-export shims for backward compatibility.
ML engine imports updated to use src/routing/ directly.
