+++ src/routing/__init__.py
@@ -1,0 +1,7 @@
+"""Shared routing infrastructure for layout engines."""
+
+from .common import DIR_MAP, DIR_VEC, DIRECTIONS, belt_entity_for_rate, machine_size, machine_tiles
+from .graph import FlowEdge, MachineNode, ProductionGraph, build_production_graph
+from .inserters import InserterAssignment, InsertionPlan, assign_inserter_positions, build_inserter_entities
+from .orchestrate import build_layout
+from .router import RoutingResult, route_connections
